+++ deploy/exec/delta-neutral-oracle/deploy/delta-neutral-oracle.ts
@@ -16,7 +16,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   ░░░╚═╝░░░╚═╝░░╚═╝░░╚═╝╚═╝░░╚═╝╚═╝░░╚══╝╚═╝╚═╝░░╚══╝░╚═════╝░
   Check all variables below before execute the deployment script
   */
-
   const deployer = (await ethers.getSigners())[0];
   const config = ConfigEntity.getConfig();
   const CHAINLINK_ADDRESS = config.Oracle.ChainLinkOracle;
