+++ package.json
@@ -190,9 +190,12 @@
     "deploy:mainnetfork:delta-neutral:deploy:oracle": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags DeltaNeutralOracle",
     "deploy:mainnetfork:delta-neutral:deploy:vault:config": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags DeltaNeutralVaultConfig",
     "deploy:mainnetfork:delta-neutral:deploy:vault": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags DeltaNeutralVault",
+    "deploy:mainnetfork:delta-neutral:configs:vault:init-positions": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags DeltaNeutralVaultInitPositions",
     "deploy:mainnetfork:delta-neutral:deploy:pcs:workers": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags DeltaNeutralPCSWorker02",
     "deploy:mainnetfork:delta-neutral:deploy:mdex:workers": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags DeltaNeutralMdexWorker02",
     "deploy:mainnetfork:delta-neutral:configs:worker:set-whitelist-caller": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags DeltaNeutralWorkerSetWhitelistCallers",
+    "deploy:mainnetfork:vault-config:config:set-whitelisted-callers": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags SetWhitelistedCallers",
+    "deploy:mainnetfork:timelock:execute": "hardhat --network mainnetfork deploy --config hardhat.config.8.10.forking.ts --no-compile --reset --tags TimeLockExecution",
     "mainnet": "hardhat console --no-compile --network mainnet",
     "ethereum": "hardhat console --no-compile --network ethereum",
     "test:case": "mocha --require hardhat/register --extension ts -g",
